requirements.txt
docx-mailmerge2==0.6.2
Faker==16.6.0
lxml==4.9.2
numpy==1.24.1
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
